Sort material data 2

+++ src/main/resources/db/migration/V2__create_components_tables.sql
@@ -48,7 +48,7 @@ CREATE TABLE graphics_cards (
     name TEXT,
     price DOUBLE PRECISION,
     chipset TEXT,
-    memory INT,
+    memory DOUBLE PRECISION,
     core_clock INT,
     boost_clock INT,
     color TEXT,
